requirements.txt
kafka-python
pandas
faker
sqlalchemy
psycopg2-binary
